LayoutTests/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-wheel-event-expected.txt

Test for wheel operations for <input type=datetime-local>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial value is 1999-01-13T00:00. We'll wheel up by 1:
PASS input.value is "1999-02-13T00:00"
Wheel up by 100:
PASS input.value is "1999-03-13T00:00"
Wheel down by 1:
PASS input.value is "1999-02-13T00:00"
Wheel down by 256:
PASS input.value is "1999-01-13T00:00"
Disabled input element:
PASS input.value is "1999-01-13T00:00"
Read-only input element:
PASS input.value is "1999-01-13T00:00"
No focus:
PASS input.value is "1999-01-13T00:00"
PASS successfullyParsed is true

TEST COMPLETE
